semi-structure-rag/requirements.txt
langchain
unstructured[all-docs]
pydantic
lxml
langchainhub
openai
chromadb
tiktoken
python-dotenv
